ProjectSupervisor

Documentation for ProjectSupervisor.

ProjectSupervisor.Supervisor

ProjectSupervisor.fetch

ProjectSupervisor.generate_or_load

ProjectSupervisor.get_gitrepo

ProjectSupervisor.get_savename

ProjectSupervisor.git_info

ProjectSupervisor.move_to_datadir

ProjectSupervisor.set_datadir

ProjectSupervisor.store_data

ProjectSupervisor.store_data

ProjectSupervisor.to_df

ProjectSupervisor.update_metadata!

ProjectSupervisor.Supervisor — Type

Supervisor

Keeps an eye on your project. Storing metadata on what gets saved where.

ProjectSupervisor.fetch — Method

fetch(sup::Supervisor; filter_criteria...)

Fetch a subset of the data stored in the metadata based on some criteria to filter which entries to select.

ProjectSupervisor.generate_or_load — Method

generateorload( fn, sup::Supervisor, paths_elements...; name=nothing, fmt=nothing, kwargs... )

If the file exists, load it. Otherwise, generate it using fn and save it.

ProjectSupervisor.get_gitrepo — Method

Find a folder that is a Git repo. Either the project or one of its parents.

ProjectSupervisor.get_savename — Method

getsavename(sup, name, fmt, pathselements...; )

Get the path where some data will be saved.

ProjectSupervisor.git_info — Method

Get info for current git repository (e.g. commit)

ProjectSupervisor.move_to_datadir — Method

Move datadir to a subfolder

ProjectSupervisor.set_datadir — Method

Move datadir to another folder.

ProjectSupervisor.store_data — Method

storedata( sup::Supervisor, data; pathselements...; name=nothing, fmt=nothing, kwargs... )

Store some data in the supervisor's data folder.

ProjectSupervisor.store_data — Method

store_data( sup::Supervisor, savename, data; metadata::AbstractDict = Dict("tag"=>"notag"), )

Store some data in the supervisor's data folder.

ProjectSupervisor.to_df — Method

Collate all metadata keys for all entries and form a dataframe with all metadata

ProjectSupervisor.update_metadata! — Method

update_metadata!(metadata::AbstractDict)::AbstractDict

Remove obsolete entries pointing to files that no longer exist.